可添加商品

+++ model/product.go
@@ -6,7 +6,7 @@ import "gorm.io/gorm"
 type Product struct {
 	gorm.Model
 	NickName    string `json:"nick_name"`   //商品名称
-	Price       int    `json:"price"`       //商品价格
+	Price       int16  `json:"price"`       //商品价格
 	Description string `json:"description"` //商品描述信息
 	Stock       int16  `json:"stock"`       //库存数量
 }
